+++ rlberry/envs/gym_make.py
@@ -1,9 +1,12 @@
 import gymnasium as gym
 
-# from gymnasium.wrappers import StepAPICompatibility
-
 from rlberry.envs.basewrapper import Wrapper
 import numpy as np
+from numpy import ndarray
+
+
+# VERSION_ORIGINE = True
+VERSION_ORIGINE = False
 
 
 def gym_make(id, wrap_spaces=False, **kwargs):
@@ -36,102 +39,122 @@ def gym_make(id, wrap_spaces=False, **kwargs):
     return Wrapper(env, wrap_spaces=wrap_spaces)
 
 
-def atari_make(id, scalarize=False, **kwargs):
+def atari_make(id, seed=None, **kwargs):
     """
-    Adaptator to work with 'make_atari_env' in stableBaselines.
-    WARNING "work in progress" : For the moment, it can't handle VecEnv, it uses only 1 env. (scalarize is forced to True)
-    (TODO PR : https://github.com/rlberry-py/rlberry/pull/285)
+    Adaptator to manage Atari Env
 
     Parameters
     ----------
     id : str
         Environment id.
-    scalarize : bool, default = False
-        If true, add a wrapper for stable_baselines VecEnv, so that they accept non-vectorized actions,
-    and return non-vectorized states. (use only the first env from VecEnv)
     **kwargs
         Optional arguments to configure the environment.
-
+        (render_mode, n_frame_stack, and other arguments for StableBaselines's make_atari_env : https://stable-baselines3.readthedocs.io/en/master/common/env_util.html#stable_baselines3.common.env_util.make_atari_env )
+    Returns
+    -------
+    Atari env with wrapper to be used as Gymnasium env.
     Examples
     --------
     >>> from rlberry.envs.gym_make import atari_make
     >>> env_ctor = atari_make
-    >>> env_kwargs = {"id": "ALE/Freeway-v5", "n_envs":1, "atari_wrappers_dict":dict(terminal_on_life_loss=False)}
+    >>> env_kwargs = {"id": "ALE/Freeway-v5", "atari_wrappers_dict":dict(terminal_on_life_loss=False),"n_frame_stack":5}}
     >>> env = env_ctor(**env_kwargs)
     """
 
-    from stable_baselines3.common.env_util import make_atari_env
-    from stable_baselines3.common.vec_env import VecFrameStack
+    from stable_baselines3.common.atari_wrappers import (  # isort:skip
+        ClipRewardEnv,
+        EpisodicLifeEnv,
+        FireResetEnv,
+        MaxAndSkipEnv,
+        NoopResetEnv,
+        NoopResetEnv,
+        StickyActionEnv,
+    )
 
-    # #uncomment when rlberry will manage vectorized env
-    # if scalarize is None:
-    #     if "n_envs" in kwargs.keys() and int(kwargs["n_envs"])>1:
-    #         scalarize = False
-    #     else:
-    #         scalarize = True
+    from stable_baselines3.common.monitor import Monitor
 
-    scalarize = True  # TODO : to remove with th PR :[WIP] Atari part2   (https://github.com/rlberry-py/rlberry/pull/285)
+    # Default values for Atari_SB3_wrappers
+    noop_max = 30
+    frame_skip = 4
+    screen_size = 84
+    terminal_on_life_loss = False  # different from SB3 : some errors with the "terminal_on_life_loss" wrapper : The 'false reset' can lead to make a step on a 'done' environment, then a crash.
+    clip_reward = True
+    action_repeat_probability = 0.0
 
-    if "atari_wrappers_dict" in kwargs.keys():
-        atari_wrappers_dict = kwargs.pop("atari_wrappers_dict")
-    else:
-        atari_wrappers_dict = dict(
-            terminal_on_life_loss=False
-        )  # hack, some errors with the "terminal_on_life_loss" wrapper : The 'false reset' can lead to make a step on a 'done' environment, then a crash.
+    if "atari_SB3_wrappers_dict" in kwargs.keys():
+        atari_wrappers_dict = kwargs.pop("atari_SB3_wrappers_dict")
+        if "noop_max" in atari_wrappers_dict.keys():
+            noop_max = atari_wrappers_dict["noop_max"]
+        if "frame_skip" in atari_wrappers_dict.keys():
+            frame_skip = atari_wrappers_dict["frame_skip"]
+        if "screen_size" in atari_wrappers_dict.keys():
+            screen_size = atari_wrappers_dict["screen_size"]
+        if "terminal_on_life_loss" in atari_wrappers_dict.keys():
+            terminal_on_life_loss = atari_wrappers_dict["terminal_on_life_loss"]
+        if "clip_reward" in atari_wrappers_dict.keys():
+            clip_reward = atari_wrappers_dict["clip_reward"]
+        if "action_repeat_probability" in atari_wrappers_dict.keys():
+            action_repeat_probability = atari_wrappers_dict["action_repeat_probability"]
 
     render_mode = None
     if "render_mode" in kwargs.keys():
         render_mode = kwargs["render_mode"]
         kwargs.pop("render_mode", None)
 
-    env = make_atari_env(env_id=id, wrapper_kwargs=atari_wrappers_dict, **kwargs)
+    if "n_frame_stack" in kwargs.keys():
+        n_frame_stack = kwargs.pop("n_frame_stack")
+    else:
+        n_frame_stack = 4
 
-    env = VecFrameStack(env, n_stack=4)
-    env = SB3_AtariImageToPyTorch(env)
-    if scalarize:
-        from rlberry.wrappers.scalarize import ScalarizeEnvWrapper
+    env = gym.make(id, render_mode=render_mode)
+    env = Wrapper(env)
+    env = Monitor(env)
 
-        env = ScalarizeEnvWrapper(env)
+    if action_repeat_probability > 0.0:
+        env = StickyActionEnv(env, action_repeat_probability)
+    if noop_max > 0:
+        env = NoopResetEnv(env, noop_max=noop_max)
+    if frame_skip > 1:
+        env = MaxAndSkipEnv(env, skip=frame_skip)
+    if terminal_on_life_loss:
+        env = EpisodicLifeEnv(env)
+    if "FIRE" in env.unwrapped.get_action_meanings():
+        env = FireResetEnv(env)
+    if clip_reward:
+        env = ClipRewardEnv(env)
+    env = gym.wrappers.ResizeObservation(env, (screen_size, screen_size))
+    env = gym.wrappers.GrayScaleObservation(env)
+    env = gym.wrappers.FrameStack(env, n_frame_stack)
+    if seed:
+        env.seed(seed)
+        env.action_space.seed(seed)
+        env.observation_space.seed(seed)
+
+    env = CompatibleWrapper(env)  # Wrapper to make it compatible with rlberry
 
     env.render_mode = render_mode
-
     return env
 
 
-class SB3_AtariImageToPyTorch(Wrapper):
-    """
-    transform the observations shape.
-    from: n_env, height, width, chan
-    to: n_env, chan, width, height
-
-    WARNING : Check the Reset and Step format :
-    https://github.com/DLR-RM/stable-baselines3/pull/1327/files#diff-a0b0c17357564df74e097f3094a5478e9b28b2af9dfdab2a91e60b6dbe174092
-
-    """
-
+class CompatibleWrapper(Wrapper):
     def __init__(self, env):
-        super(SB3_AtariImageToPyTorch, self).__init__(env)
-        old_shape = self.observation_space.shape
-        new_shape = (old_shape[-1], old_shape[0], old_shape[1])
-        self.observation_space = gym.spaces.Box(
-            low=0.0, high=1.0, shape=new_shape, dtype=np.float32
-        )
+        super(CompatibleWrapper, self).__init__(env)
+        self.render_mode = None
 
-    def observation(self, observation):
-        return np.transpose(observation, (0, 3, 2, 1))  # transform
+    def step(self, action):
+        if type(action) is ndarray and action.size == 1:
+            action = action[0]
 
-    def reset(self):
-        obs = self.env.reset()
-        infos = self.env.reset_infos
-        return self.observation(obs), infos
-
-    def step(self, actions):
-        next_observations, rewards, done, infos = self.env.step(actions)
-        # return self.observation(next_observations), rewards, done, [d["TimeLimit.truncated"] for d in infos], infos
+        next_observations, rewards, terminated, truncated, infos = self.env.step(action)
         return (
-            self.observation(next_observations),
+            np.array(next_observations),
             rewards,
-            done,
-            [None] * self.env.num_envs,
+            terminated,
+            truncated,
             infos,
         )
+
+    def reset(self, seed=None, options=None):
+        obs, infos = self.env.reset(seed=seed, options=options)
+
+        return np.array(obs), infos